memory sizes are stored in size_t variables instead of ints. Removes remaining 32/64 signed/unsigned warnings
git-svn-id: http://wnd-charm.googlecode.com/svn/pychrm/branches/shared_ImageMatrix@531 d860a611-dd6a-fbde-3e4e-7fa7191099b5

+++ src/b64/cencode.c
@@ -26,7 +26,7 @@ char base64_encode_value(char value_in)
 	return encoding[(int)value_in];
 }
 
-int base64_encode_block(const char* plaintext_in, int length_in, char* code_out, base64_encodestate* state_in)
+size_t base64_encode_block(const char* plaintext_in, size_t length_in, char* code_out, base64_encodestate* state_in)
 {
 	const char* plainchar = plaintext_in;
 	const char* const plaintextend = plaintext_in + length_in;
@@ -87,7 +87,7 @@ int base64_encode_block(const char* plaintext_in, int length_in, char* code_out,
 	return codechar - code_out;
 }
 
-int base64_encode_blockend(char* code_out, base64_encodestate* state_in)
+size_t base64_encode_blockend(char* code_out, base64_encodestate* state_in)
 {
 	char* codechar = code_out;
