+++ src/b64/cencode.h
@@ -8,6 +8,9 @@ For details, see http://sourceforge.net/projects/libb64
 #ifndef BASE64_CENCODE_H
 #define BASE64_CENCODE_H
 
+#include <stdlib.h> // for size_t
+
+
 typedef enum
 {
 	step_A, step_B, step_C
@@ -26,8 +29,8 @@ void base64_init_encodestate(base64_encodestate* state_in);
 
 char base64_encode_value(char value_in);
 
-int base64_encode_block(const char* plaintext_in, int length_in, char* code_out, base64_encodestate* state_in);
+size_t base64_encode_block(const char* plaintext_in, size_t length_in, char* code_out, base64_encodestate* state_in);
 
-int base64_encode_blockend(char* code_out, base64_encodestate* state_in);
+size_t base64_encode_blockend(char* code_out, base64_encodestate* state_in);
 
 #endif /* BASE64_CENCODE_H */
